segmented-switch: refactor interface to have title
Currently, SegmentedSwitch component allow to pass an array of strings
to prop `items` and a string to prop `selected` for the current option.
The items strings are used as titles for the options, which could be
strange in situations where your state model differs the title you'd
like to display.

This refactors the options/selected value to this interface:

  <SegmentedSwitch
    options={[
      {
        title: 'Portuguese',
        value: 'pt-BR',
      },
      {
        title: 'English',
        value: 'en-US',
      },
    ]}
    value="pt-BR"
  />